requirements.txt
asgiref==3.3.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-filter==21.1
django-jsonfield==1.4.1
djangorestframework==3.11.0
idna==2.9
jsonfield==3.1.0
pytz==2020.1
requests==2.23.0
six==1.16.0
sqlparse==0.3.1
unittest-xml-reporting==3.0.2
urllib3==1.25.9
